defect #820004 :  error notification on permissions insufficient from Jenkins plugin is missing info in Octane,defect #616017 : can "run successfully" disabled job from Octane
